third_party/WebKit/LayoutTests/fast/canvas-api/canvas-2d-imageData-create-nonfinite.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<canvas id="canvas" width=600 height=300>
<script>
test(function(t) {
    var canvas = document.getElementById('canvas');
    var ctx = canvas.getContext('2d');

    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.createImageData(Infinity, Infinity);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.createImageData(Infinity, 10);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.createImageData(-Infinity, 10);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.createImageData(10, Infinity);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.createImageData(10, -Infinity);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.createImageData(NaN, 10);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.createImageData(10, NaN);
    });
}, 'Test the argument bounds of canvas createImageData.');
</script>
